Increase Quantity property type from byte to int
Changed the Quantity property in both ProductItem and ShipmentItem
classes from byte to int to support larger quantity values. This
removes the previous limitation of a maximum value of 255.

+++ DataFormulaLibrary/Models/Shipping/ShipmentItem.cs
@@ -9,7 +9,7 @@ public class ShipmentItem
 	public int Id { get; set; } = 0;
 	public int ShipmentId { get; set; } = 0;
 	public int OrderItemId { get; set; } = 0;
-	public byte Quantity { get; set; } = 0;
+	public int Quantity { get; set; } = 0;
 	public float Weight { get; set; } = 0.0f;
 	public int WarehouseId { get; set; } = 0;
 }
